WW-3714 Deprecate and migrate Action, Interceptor, Result
