+++ core/src/main/java/com/opensymphony/xwork2/Result.java
@@ -18,33 +18,11 @@
  */
 package com.opensymphony.xwork2;
 
-import java.io.Serializable;
-
 /**
- * All results (except for <code>Action.NONE</code>) of an {@link Action} are mapped to a View implementation.
+ * {@inheritDoc}
  *
- * <p>
- * Examples of Views might be:
- * </p>
- *
- * <ul>
- * <li>SwingPanelView - pops up a new Swing panel</li>
- * <li>ActionChainView - executes another action</li>
- * <li>SerlvetRedirectView - redirects the HTTP response to a URL</li>
- * <li>ServletDispatcherView - dispatches the HTTP response to a URL</li>
- * </ul>
- *
- * @author plightbo
+ * @deprecated since 6.7.0, use {@link org.apache.struts2.Result} instead.
  */
-public interface Result extends Serializable {
-
-    /**
-     * Represents a generic interface for all action execution results.
-     * Whether that be displaying a webpage, generating an email, sending a JMS message, etc.
-     *
-     * @param invocation  the invocation context.
-     * @throws Exception can be thrown.
-     */
-    void execute(ActionInvocation invocation) throws Exception;
-
+@Deprecated
+public interface Result extends org.apache.struts2.Result {
 }
